update install file

+++ install.sh
@@ -1,38 +1,46 @@
 #!/bin/bash
 
+exists()
+{
+  command -v "$1" >/dev/null 2>&1
+}
+
 # Installing all Deps and Configs
-
-#### Deps ####
-
-echo "----- deps -----"
-
-## Fish
-if brew ls --versions fish > /dev/null; then
-  echo "fish exists"
-else
-  echo "installing fish"
-  brew install fish
-fi
+echo "----------------"
+echo "Dependencies"
+echo "----------------"
 
 ## Git
-if git --version > /dev/null; then
+if exists git; then
   echo "git exists"
 else 
   echo "installing git"
   brew install git
 fi
 
-## Omf (pkg manager for fish)
-if [ ! -d ~/.config/omf ]; then
-  curl -sL https://git.io/fisher | source
+## Fish
+if exists fish; then
+  echo "fish exists"
+else
+  echo "Installing Fish"
+  brew install fish
   sudo bash -c 'echo /usr/local/bin/fish >> /etc/shells'
   chsh -s /usr/local/bin/fish
   
+fi
+
+## Fisher
+if [ ! -d /.config/fish/functions/fisher.fish ]; then
+  echo "fisher exists"
+else
+  brew tap bbatsche/fisher
+  brew install fisherman
+
   fisher install matchai/spacefish
 fi
 
 ## Tmux
-if brew ls --versions tmux > /dev/null; then
+if exists tmux; then
   echo "tmux exists"
 else
   echo "installing tmux"
@@ -40,7 +48,7 @@ else
 fi
 
 ## NeoVim
-if brew ls --versions neovim > /dev/null; then
+if exists nvim; then
   echo "neo vim exists"
 else
   echo "installing nvim"
@@ -50,47 +58,55 @@ fi
 ## Alacritty
 if [ ! -d /Applications/Alacritty.app ]; then
   echo "moving alacritty app to applications"
+  // TODO: copy app to applications
 else
   echo "alacritty app exists"
 fi
 
 ## Hammerspoon
-if brew info --cask hammerspoon > /dev/null; then
-  echo "hammerspoon exists"
-else 
+if [ ! -d /Applications/Hammerspoon.app ]; then
+  echo "installing hammerspoon"
   brew install hammerspoon --cask
+else 
+  echo "hammerspoon exists" 
 fi
 
-
-#### Configs ####
-echo "----- configs -----"
+echo "----------------"
+echo "Configs"
+echo "----------------"
 
 ## Hammerspoon
-echo "copying hammerspoon configs"
-
 if [ ! -d ~/.hammerspoon ]; then
+  echo "copying hammerspoon configs"
   mkdir ~/.hammerspoon
+  yes | cp -rf ./hammerspoon/* ~/.hammerspoon/
+else 
+  echo "hammerspoon configs exist -> update" 
+  yes | cp -rf ./hammerspoon/* ~/.hammerspoon/
 fi
-
-yes | cp -rf ./hammerspoon/* ~/.hammerspoon/
 
 ## Alacritty
-echo "copying alacritty configs"
-
 if [ ! -d ~/.config/alacritty ]; then
-  mkdir ~/.config alacritty
+  echo "copying alacritty configs"
+  cd ~/.config && mkdir alacritty
+  yes | cp -rf ./alacritty/alacritty.yml ~/.config/alacritty
+  yes | cp -rf ./alacritty/dracula.yml ~/.config/alacritty
+else 
+  echo "alacritty configs exist -> update" 
+  yes | cp -rf ./alacritty/alacritty.yml ~/.config/alacritty
+  yes | cp -rf ./alacritty/dracula.yml ~/.config/alacritty
 fi
 
-# copy config
-yes | cp -rf ./alacritty/alacritty.yml ~/.config/alacritty
-
-# copy theme file
-yes | cp -rf ./alacritty/dracula.yml ~/.config/alacritty
-
 ## Tmux
+if [ ! -f ~/.tmux.conf ]; then
+  echo "copying tmux configs"
+  yes | cp -rf ./tmux/.tmux.conf ~/
+  tmux source-file ~/.tmux.conf
+else 
+  echo "tmux configs exist -> update" 
+  yes | cp -rf ./tmux/.tmux.conf ~/
+  tmux source-file ~/.tmux.conf
+fi
 
-echo "copying tmux configs"
 
-yes | cp -rf ./tmux/.tmux.conf ~/
-tmux source-file ~/.tmux.conf
